refactor: move user accounts into a platform app, without touching a table
A user's profile, Patreon status and chosen badge belong to the person,
not to the edition they happen to play. UserProfile and PatreonStatus move
from n23.core to a new platform app, gyrinx.accounts (label `accounts`).

This is a real app-label change, which normally means renaming tables. It
doesn't here: both tables are pinned to their existing names —
db_table = "core_userprofile" and HistoricalRecords(table_name=
"core_historicaluserprofile") — and both migrations are wrapped in
SeparateDatabaseAndState with no database_operations. Django's model state
moves; the 2,803 production rows do not.

The ContentType rows are updated in place from core to accounts rather
than left to post_migrate. Keeping the same content_type_id means the four
auth_permission rows follow automatically, and any GenericForeignKey
pointing at a profile stays valid. Checked against production first: zero
Event rows point at userprofile, so nothing needed repointing.

Following the model, because all three are platform concerns that were
only edition-located: the User/Group/EmailAddress/UserProfile admin (370
lines, zero edition imports) to gyrinx/accounts/admin.py;
BadgeSelectionForm to gyrinx.account_forms, which was waiting on exactly
this; and the admin's tests to gyrinx/tests/.

Verified on a populated database: 6 rows before and after, ContentTypes
130/131 keeping their ids, tables still named core_*, and zero accounts_*
tables created. makemigrations --check clean, 4166 tests pass.

Refs #2093
