+++ bandit/bandit-baseline.txt
@@ -1,4 +1,4 @@
-Run started:2026-08-04 22:06:42.593548+00:00
+Run started:2026-08-04 22:16:35.182848+00:00
 
 Test results:
 >> Issue: [B703:django_mark_safe] Potential XSS on mark_safe function.
@@ -23,14 +23,14 @@ Test results:
    Severity: Low   Confidence: Medium
    CWE: CWE-259 (https://cwe.mitre.org/data/definitions/259.html)
    More Info: https://bandit.readthedocs.io/en/1.9.4/plugins/b105_hardcoded_password_string.html
-   Location: ./gyrinx/settings.py:301:4
-300	ACCOUNT_FORMS = {
-301	    "reset_password": "gyrinx.account_forms.ResetPasswordForm",
-302	    "login": "gyrinx.account_forms.LoginForm",
-303	    "signup": "gyrinx.account_forms.SignupForm",
-304	}
-305	# Custom setting to (dis)allow signups
-306	ACCOUNT_ALLOW_SIGNUPS = os.getenv("ACCOUNT_ALLOW_SIGNUPS", "True") == "True"
+   Location: ./gyrinx/settings.py:302:4
+301	ACCOUNT_FORMS = {
+302	    "reset_password": "gyrinx.account_forms.ResetPasswordForm",
+303	    "login": "gyrinx.account_forms.LoginForm",
+304	    "signup": "gyrinx.account_forms.SignupForm",
+305	}
+306	# Custom setting to (dis)allow signups
+307	ACCOUNT_ALLOW_SIGNUPS = os.getenv("ACCOUNT_ALLOW_SIGNUPS", "True") == "True"
 
 --------------------------------------------------
 >> Issue: [B324:hashlib] Use of weak MD5 hash for security. Consider usedforsecurity=False
@@ -340,7 +340,7 @@ Test results:
 --------------------------------------------------
 
 Code scanned:
-	Total lines of code: 71188
+	Total lines of code: 71196
 	Total lines skipped (#nosec): 4
 	Total potential issues skipped due to specifically being disabled (e.g., #nosec BXXX): 56
