+++ gyrinx/accounts/__init__.py
@@ -1,0 +1,5 @@
+"""User accounts and supporter status — platform, not edition.
+
+A user's profile, Patreon membership and chosen badge belong to the person,
+not to whichever game edition they happen to be playing.
+"""
