+++ gyrinx/accounts/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AccountsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "gyrinx.accounts"
+    label = "accounts"
