+++ gyrinx/api/patreon.py
@@ -3,7 +3,7 @@ import logging
 from allauth.account.models import EmailAddress
 
 from gyrinx.badges import rank_for_tier_title
-from n23.core.models.auth import PatreonStatus, UserProfile
+from gyrinx.accounts.models import PatreonStatus, UserProfile
 
 logger = logging.getLogger(__name__)
