+++ n23/core/admin/__init__.py
@@ -1,5 +1,4 @@
 from .action import *  # noqa: F403
-from .auth import *  # noqa: F403
 from .backfill import *  # noqa: F403
 from .campaign import *  # noqa: F403
 from .events import *  # noqa: F403
