+++ n23/core/models/__init__.py
@@ -1,5 +1,4 @@
 from .action import *  # noqa: F403
-from .auth import *  # noqa: F403
 from .backfill import *  # noqa: F403
 from .battle import *  # noqa: F403
 from .campaign import *  # noqa: F403
