+++ n23/core/views/user.py
@@ -7,10 +7,10 @@ from django.db.models import Count, Q
 from django.shortcuts import get_object_or_404, redirect, render
 
 from gyrinx.badges import HIDE_BADGE
-from n23.core.forms import BadgeSelectionForm
+from gyrinx.account_forms import BadgeSelectionForm
 from gyrinx.account_forms import UsernameChangeForm
 from gyrinx.impersonation import can_impersonate_target
-from n23.core.models.auth import UserProfile
+from gyrinx.accounts.models import UserProfile
 from n23.core.models.campaign import Campaign
 from n23.core.models.events import EventNoun, EventVerb, log_event
 from n23.core.models.list import List
